+++ src/kernel/kernel.cc
@@ -1,7 +1,7 @@
 #include <sys/kernel/kernel.hh>
 
-extern "C" [[noreturn]] void sys_kernel_entry() noexcept {
-    sys::kernel::start();
+extern "C" [[noreturn]] void sys_kernel_entry(sys::uintptr_t firmware_data) noexcept {
+    sys::kernel::start(firmware_data);
 }
 
 extern "C" [[noreturn]] void sys_kernel_secondary_entry() noexcept {